% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/getpromise.R
\name{arg}
\alias{arg}
\alias{arg_}
\title{Capture an argument by name from the present environment.}
\usage{
arg(sym, env = arg_env_(quote(sym), environment()))

arg_(name, env = arg_env(name, environment()))
}
\arguments{
\item{sym}{the name to look up. For \code{arg} this is a symbol
and not evaluated; for \code{arg_} this is a symbol or
character.}

\item{env}{The environment to look in.}
}
\value{
A \link{quo} object.
}
\description{
Capture an argument by name from the present environment.
}
