% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/getpromise.R
\name{arg_env_}
\alias{arg_env_}
\title{_arg_env(x, env) is the normally evaluating version of arg_env; you
supply a vatiable name and an environment to look in.}
\usage{
arg_env_(name, env = arg_env_(quote(name), environment()))
}
\description{
_arg_env(x, env) is the normally evaluating version of arg_env; you
supply a vatiable name and an environment to look in.
}
