% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/getpromise.R
\name{args}
\alias{args}
\alias{args_}
\title{Capture named arguments from the present environment, returning a dots list.}
\usage{
args(...)

args_(syms, envs)
}
\arguments{
\item{...}{Variable names (unevaluated). Arguments may be named; these names
determine the names on the dots list (and not the variable names).}

\item{envs}{An environment, or list of environments, to look for
the bindings in.}

\item{names}{A character vector or list of names.}
}
\value{
a \code{\link{dots}} object containing the promises that are bound to
those variables in the calling environment.
}
\description{
Capture named arguments from the present environment, returning a dots list.

\code{args_} is a normally evaluating version of \code{args}.
}
\note{
If no argument names are given, none are used.
}
\seealso{
dots get_dots
}
