% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/getpromise.R
\name{arg_promise}
\alias{arg_promise}
\alias{arg_promise_}
\title{Fetch a single promise by name from an environment.}
\usage{
arg_promise(name)

arg_promise_(name, env)
}
\arguments{
\item{name}{the name to look up. For \code{arg_promise} this is a symbol
and not evaluated; for \code{arg_promise_} this is a symbol or
character.}

\item{env}{The environment to look in.}
}
\value{
A promise object.
}
\description{
Fetch a single promise by name from an environment.

..
}
