% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/conversions.R
\name{as.dots.environment}
\alias{as.dots.environment}
\alias{dots2env}
\title{Capture promises present in an environment and construct a dots object.}
\usage{
\method{as.dots}{environment}(x)
}
\description{
All bindings in the environment (but not those from its parents) will be
copied into a new \code{\dots} list. Bindings that are promises
will be added to the \dots list without forcing, while other
bindings will be wrapped in an already-forced promise.  If \code{...} is
bound in the environment, all bindings it contains will be added to
the \dots list. The output will not be in any particular order.
}
