% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/conversions.R
\name{as.dots.exprs}
\alias{as.dots.exprs}
\title{Convert a list of expressions into a dots object.}
\usage{
\method{as.dots}{exprs}(exprs, env = arg_env(exprs))
}
\arguments{
\item{exprs}{a list of expressions}

\item{env}{An environment or list of environments.}
}
\description{
Convert a list of expressions into a dots object.
}
