% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/conversions.R
\name{as.dots}
\alias{as.dots}
\alias{as.dots.dots}
\alias{as.dots.quotation}
\alias{as.dots.list}
\alias{as.dots.environment}
\alias{as.dots.lazy_dots}
\alias{as.dots.default}
\title{Convert items into quotations or dots.}
\usage{
as.dots(x)

\method{as.dots}{dots}(x)

\method{as.dots}{quotation}(x)

\method{as.dots}{list}(x)

\method{as.dots}{environment}(x)

\method{as.dots}{lazy_dots}(x)

\method{as.dots}{default}(x)
}
\arguments{
\item{x}{a vector or list.}
}
\value{
An object of class \code{\dots}.
}
\description{
\code{as.dots} is a generic function for converting data into \link{dots}.

\code{as.dots.environment} is a synonym for \code{\link{env2dots}}.
}
\seealso{
env2dots
rdname dots2env
}
