% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/quo.R
\name{as.quo.literal}
\alias{as.quo.literal}
\title{Put data values into quotations.}
\usage{
\method{as.quo}{literal}(x)
}
\value{
as.quo.literal
}
\description{
\code{as.quo.literal(x)} creates a forced quotation with the given value.
}
