% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/caller.R
\name{call_from}
\alias{call_from}
\title{Call a function from a particular environment.}
\usage{
call_from(f, envir)
}
\arguments{
\item{f}{The function to call. A name or character string may be
used, which should refer to a function visible in the environment
given.}

\item{envir}{The environment to issue the call from. If \code{\link{caller}}
is called within \code{f}, it will return \code{envir}.}
}
\value{
A function wrapper. Normally one would immediately invoke it, as in
call_from(f, e)("foo", "bar", "baz").

You can also use this to insulate yourself from poorly behaved functions that
use parent.frame() when they should use arg_env(), as in \link{plyr::here}.
}
\description{
Call a function from a particular environment.
}

