% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/caller.R
\name{call_here}
\alias{call_here}
\title{Wrap a function so that it will see a particular caller or parent.frame.}
\usage{
call_here(f)
}
\description{
Writing \code{call_here(f)} is equivalent to writing
\code{call_from(f, environment())}.
}

