% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/conversions.R
\name{as.lazy_dots}
\alias{as.lazy_dots}
\alias{as.lazy_dots.dots}
\title{Compatibility conversions.}
\usage{
as.lazy_dots(x, env)

\method{as.lazy_dots}{dots}(x, env = "ignored")
}
\arguments{
\item{x}{A \link[lazyeval:lazy_dots]{lazyeval::lazy_dots} object.}

\item{env}{See [[lazyeval::as.lazy_dots]]

[[lazyeval::as.lazy_dots]: R:[lazyeval::as.lazy_dots}
}
\description{
Convert quotations and dot lists to the representations used
by other packages.
}
\seealso{
as.dots
}
