% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dots.R
\name{dotc}
\alias{dotc}
\title{dotc casts each of its arguments to a dotlist, and concatenates
the lot, returning a dotlist}
\usage{
dotc(...)
}
\arguments{
\item{...}{Things that can be converted to args or dots.}
}
\description{
dotc casts each of its arguments to a dotlist, and concatenates
the lot, returning a dotlist
}
\seealso{
quo
}
