% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dots.R
\name{dots_environments}
\alias{dots_environments}
\alias{environments}
\alias{environments<-}
\alias{environments<-.dots}
\title{Extract or manipulate environments contained in dots lists.}
\usage{
dots_environments(...)

environments(x)

environments(x) <- value

\method{environments}{dots}(x) <- value
}
\arguments{
\item{...}{Any arguments.}

\item{x}{a \{code{\link{dots}} object.}

\item{value}{A new list of environments to apply.}
}
\value{
A named list of environment objects. The mutator
\code{environments<-} constructs a new list of unevaluated promises
with the same expressions but different environments.
}
\description{
\code{environments} works on a dots list created by \{code{\link{dots}} w,
while \code{dots_environments} works on arguments directly passed in.
}

