% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dots.R
\name{dots_unpack}
\alias{dots_unpack}
\alias{unpack}
\alias{unpack}
\title{Show information about a \dots object.}
\usage{
dots_unpack(...)

unpack(x)
}
\arguments{
\item{...}{Any number of arguments. Usually, you will pass in the
... from the body of a function,
e.g. \code{dots_unpack(...)}.}

\item{x}{A \code{\link{dots}} object.}
}
\value{
A data frame, with one row for each element of \code{\...},
and columns: \describe{ \item{"name"}{The name of each argument,
or "" if present.}  \item{"envir"}{The enviroment the promise
came from.}  \item{"expr"}{The expression attached to the
promise. If the promise has been evaluated, this will be NULL.}
\item{"value"}{The value attached to the promise. If the promise
has not been evaluated, this will be NULL.}}
}
\description{
\code{unpack} transforms the contents of a \dots object, into a data
frame with columns "name", "expr", "envir" and "value".
}
\note{
There are some issues with R printing data frames containing
lists of language objects (and more problems when working with
"missing value" objects.) Therefore this sets the class on the
columns to one that has a special as.character method.
}
\seealso{
dots_names dots_missing dots_exprs dots
}
