% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dots.R
\name{envs}
\alias{envs}
\alias{envs<-}
\title{Extract or update environments contained in dots lists.}
\usage{
envs(x)

envs(x) <- value
}
\arguments{
\item{x}{a \{code{\link{dots}} object.}

\item{value}{An environment or list of environments.}
}
\value{
A named list of environment objects. The mutator
  \code{envs<-} constructs a new list of unevaluated promises with
  the same expressions but different environments.
}
\description{
\code{envs(dots(...)} extracts a list of environments from a
\code(link(dots)) object.
}
