% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dots.R
\name{exprs}
\alias{exprs}
\title{Extract unevaluated expressions.}
\usage{
exprs(x)
}
\arguments{
\item{x}{A dots object (see \code{\link{dots}}).}
}
\value{
A named list of expressions. The mutator \code{exprs<-}
  constructs a new dots object with the new expressions.
}
\description{
\code{exprs(dots(...))} extracts a list of expressions, one per element
of a \code{\link{dots}} object.
}
