% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/getpromise.R
\name{find}
\alias{find}
\alias{find_}
\title{Find the enclosing environment that defines a name.}
\usage{
find(x, mode = "any")

find_(x, env, mode = "any", ifnotfound = stop("Binding ", x, " not found"))
}
\arguments{
\item{x}{A name.}

\item{mode}{Either "any" or "function".}
}
\description{
Find the enclosing environment that defines a name.

\code{find_} is the normally evaluating version, which takes an environment.
}
