% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/format.R
\name{format.dots}
\alias{format.dots}
\title{Format a dots object for printing.}
\usage{
\method{format}{dots}(x, compact = FALSE, show.environments = !compact,
  show.expressions = !compact, width = 30, ...)
}
\arguments{
\item{x}{A dots object.}

\item{compact}{Implies \code{show.environments=FALSE} and
\code{show.expressions=FALSE}.}

\item{show.environments}{Whether to show environments for unevaluated
promises.}

\item{show.expressions}{Whether to show expressions for evaluated promises.}

\item{...}{Further arguments passed to \link{format} methods.}
}
\description{
Constructs a string representation of a dots object. In this representation
an unevaluated promise is printed as "\code{envir ? expr}" and an evaluated
promise is shown as "\code{expr := value}".
}
