% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/format.R
\name{format.oneline}
\alias{format.oneline}
\title{The method \code{format.oneline} formats a vector or list so that each
item fits on one line. It is similar to \link{format.AsIs} but tries
harder with language objects. The "oneline" class is used by
\link{as.data.frame.dots}.}
\usage{
\method{format}{oneline}(x, max.width = 50, width = max.width, ...)
}
\arguments{
\item{x}{A vector or list.}

\item{max.width}{See \link{format}.}

\item{width}{See \link{format}.}

\item{...}{Further arguments passed to \link{format}.}
}
\description{
The method \code{format.oneline} formats a vector or list so that each
item fits on one line. It is similar to \link{format.AsIs} but tries
harder with language objects. The "oneline" class is used by
\link{as.data.frame.dots}.
}
