% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/format.R
\name{format.dots}
\alias{format.dots}
\alias{format.quotation}
\alias{print.dots}
\alias{print.quotation}
\title{Formatting methods for dots and quotations.}
\usage{
\method{format}{dots}(x, compact = FALSE, show.environments = !compact,
  show.expressions = !compact, width = 30, ...)

\method{format}{quotation}(x, compact = FALSE, show.environments = !compact,
  show.expressions = !compact, width = 30, ...)

\method{print}{dots}(x, ...)

\method{print}{quotation}(x, ...)
}
\arguments{
\item{x}{An object.}

\item{compact}{Implies \code{show.environments=FALSE} and
\code{show.expressions=FALSE}.}

\item{show.environments}{Whether to show environments for unforced
quotations.}

\item{show.expressions}{Whether to show expressions for forced
quotations.}

\item{width}{the maximum length to use for each element.}

\item{...}{Parameters passed to \link[base:format]{base::format}}
}
\description{
\code{format.dots} constructs a string representation of a dots
object. An un\link{forced} quotation is shown as \code{envir ? expr} and a
forced quotation is shown as \code{expr := value}.

\code{format.quotation} constructs a string representation of a
quotation object.
}
