% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/getpromise.R
\name{\%<-\%}
\alias{\%<-\%}
\title{Turn quotations into bindings in the present environment.}
\usage{
dst \%<-\% src
}
\arguments{
\item{dst}{A name,}

\item{src}{A \link{quotation} (or something that can be converted to a
quotation, like a formula).}
}
\description{
Turn quotations into bindings in the present environment.
}
\seealso{
delayedAssign
}
