% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/getpromise.R
\name{locate}
\alias{locate}
\alias{locate_}
\alias{locate_.list}
\alias{locate_.quotation}
\alias{locate_.character}
\alias{locate_.(}
\alias{locate_.name}
\title{Determine which enclosing environment defines a name.}
\usage{
locate(sym, env = arg_env_(quote(sym), environment()), mode = "any", ...)

locate_(sym, env = arg_env_(quote(sym), environment()), mode = "any", ...)

locate_.list(sym, env = arg_env_(quote(sym), environment()), mode = "any",
  ...)

locate_.quotation(sym, env = (ignored), mode = "any", ...)

locate_.character(sym, env = arg_env_(quote(sym), environment()),
  mode = "any", ...)

"locate_.("(sym, env = arg_env_(quote(sym), environment()), mode = "any",
  ...)

locate_.name(sym, env = arg_env_(quote(sym), environment()), mode = "any",
  ifnotfound = stop("Binding ", deparse(sym), " not found"))
}
\arguments{
\item{sym}{A name. For \code{locate} the argument is used literally. For
\code{locate_} it should have the value \link{name} or character.}

\item{env}{Which environment to begin searching from.}

\item{mode}{Either "any" or "function". "any" finds the lowest
enclosing environment which defines a symbol. "function" finds an
environment which defines the symbol as a function, possibly
forcing promises along the way.}

\item{ifnotfound}{What is returned if the symbol is not found. By
default an exception is raised.}

\item{mode}{Either "any" or "function".}
}
\value{
An environment object which defines \code{sym}, if one is found.
}
\description{
Determine which enclosing environment defines a name.

If sym is a list (of \link{name}s) or a \link{dots} object, \code{locate} returns a list.

When \code{sym} is a \link{quotation} or \link{dots}, the \code{env} argument is ignored.
}
\note{
If you use a literal character argument, as in \code{locate("x", environment())}, you must also provide the environment
argument. However \code{locate(x)} will work OK. See note under \link{arg}.
}
\examples{
# `locate` is useful if you want to implement something that works
# like [<<-], which updates the binding where it is found.
`<<-` <- function(lval, rval) {
 lval_ <- arg(lval_)
 rval_ <- arg(rval_)
 target.env <- locate_(expr(lval_), parent.env(env(lval_)))
 do_(quo(`<-`, target.env), lval_, rval_)
}

x <- "not this one"
local({
  x <- "this one"
  local({
    x <- "not this one"
    x <<- "this works just like builtin <<-"
  })
  print(x)
})
}
