% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/getpromise.R
\name{locate_.quotation}
\alias{locate_.quotation}
\title{The locate_ method for quotations uses the expr and environment together.}
\usage{
\method{locate_}{quotation}(sym, ..., mode = "any")
}
\description{
The locate_ method for quotations uses the expr and environment together.
}
