% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/conversions.R
\name{promise}
\alias{promise}
\alias{promise_}
\title{Create a representation of a promise.}
\usage{
promise(x)

promise_(expr, env)
}
\arguments{
\item{expr}{The expression.}

\item{env}{The environment.}
}
\value{
A representation of a promise.
}
\description{
A promise contains two data items: an expression, and an
environment.  \code{promise} captures its argument without
evaluating and returns a promise object.

\code{promise_} is normally evaluating and constructs a promise
from an expression and an environment.
}
