% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/conversions.R
\name{quo2env}
\alias{quo2env}
\title{Bind a name in an environment to a promise.}
\usage{
quo2env(q, env, name)
}
\arguments{
\item{q}{a quotation.}

\item{env}{The environment to store in.}

\item{name}{The name to use. If "" or NULL, will append to "...".}
}
\description{
Bind a name in an environment to a promise.
}
