% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dots.R
\name{set_dots}
\alias{set_dots}
\title{Set the binding of "..." in an environment.}
\usage{
set_dots(env, dots, append = FALSE)
}
\arguments{
\item{env}{The environment to update.}

\item{dots}{a \code{\link{dots}} object.}

\item{append}{Whether to append to an existing binding for "..." if
the environment has one.}
}
\value{
the updated environment, invisibly.
}
\description{
If an empty dotlist is given, the "..." binding is set to the
missing value.
}
