% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dots.R
\name{value}
\alias{value}
\title{Evaluate promises or dots.}
\usage{
value(p)
}
\arguments{
\item{p}{a promise object.}

\item{d}{a dots object.}
}
\value{
a value or list of values.
}
\description{
For promise objects, evaluates the given promise and returns the
result. For "dots" objects, evaluates each element and returns a
list of results. Unlike native R promises, the results will not be
retained.
}
