% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/caller.R
\name{with_caller}
\alias{with_caller}
\title{Wrap a function so that it will see a particular caller or parent.frame.}
\usage{
with_caller(f, envir = arg_env(f))
}
\arguments{
\item{f}{The function to call. A symbol may be given, which must be a symbol visible 
from \code{e}.}

\item{envir}{The environment to issue the call from. If \code{\link{caller}()}
is called within \code{f}, it will return \code{envir}.}
}
\description{
Wrap a function so that it will see a particular caller or parent.frame.
}

