% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/getpromise.R
\name{arg}
\alias{arg}
\alias{arg_}
\title{Fetch an argument by name from an environment.}
\usage{
arg(name)

arg_(name, env)
}
\arguments{
\item{name}{the name to look up. For \code{arg} this is a symbol
and not evaluated; for \code{arg_} this is a symbol or
character.}

\item{env}{The environment to look in.}
}
\value{
A \link{quo} object.
}
\description{
Fetch an argument by name from an environment.

..
}
