% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/getpromise.R
\name{args}
\alias{args}
\alias{args_}
\title{Fetch an argument from an environment by name.}
\usage{
args(...)

args_(names, envs, tags = names(names) \%||\% vapply(names, as.character, ""))
}
\arguments{
\item{...}{Variable names (unevaluated). Arguments may be named; these names
determine the names on the dots list (and not the variable names)}

\item{names}{A character vector or list of names.}

\item{envs}{An environment, or list of environments, to look for
the bindings in.}

\item{tags}{An optional character vector specifying output variable names.}
}
\value{
a \code{\link{dots}} object containing the promises that are bound to
those variables in the calling environment.
}
\description{
Fetch an argument from an environment by name.

\code{arg_dots_} is a normally evaluating version of args.
}
\note{
The tags on the dots object are determined by argument names;
variable names are discarded.
}
\seealso{
dots get_dots
}
