% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/conversions.R
\name{as.dots}
\alias{as.dots}
\alias{as.dots.literal}
\title{Convert a list of expressions into a \code{\dots} object (a list of
promises.)}
\usage{
as.dots(x)

\method{as.dots}{literal}(values)
}
\arguments{
\item{x}{a vector or list.}

\item{env}{The environment to attach to each expression.}

\item{x}{a list.}
}
\value{
An object of class \code{\dots}. For \code{as.dots}, the
list items are treated as data values, and create already-forced
promises. For \code{as.dots.exprs}, values are treated as the
expressions of new unforced promises.
}
\description{
Convert a list of expressions into a \code{\dots} object (a list of
promises.)

\code{as.dots.literal} turns a list of data values into a dots object
consisting of forced promises. (A forced promise records a value
and an expression, but not an environment.)
}
