% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/conversions.R
\name{as.lazy_dots.dots}
\alias{as.lazy_dots.dots}
\alias{as.lazy.quotation}
\alias{as.quo.quosure}
\alias{as.quo.formula}
\alias{as.quosure.quo}
\alias{as.quosures.dots}
\title{Compatibility conversions.}
\usage{
as.lazy_dots.dots(x, env = "ignored")

as.lazy.quotation(x, env = "ignored")

\method{as.quo}{quosure}(x)

\method{as.quo}{formula}(x)

as.quosure.quo(x)

as.quosures.dots(x)
}
\arguments{
\item{x}{an object to convert.}

\item{env}{Ignored for quotations or dots.}
}
\value{
The converted object.
}
\description{
Convert quotations and dot lists to representations used
by some other packages.
}
