% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/conversions.R
\name{as.lazy_dots}
\alias{as.lazy_dots}
\alias{as.lazy_dots.dots}
\title{Compatibility conversions.}
\usage{
as.lazy_dots(x, env)

\method{as.lazy_dots}{dots}(x, env = "ignored")
}
\arguments{
\item{x}{a \link{dots} object.}

\item{env}{See \link[lazyeval:as.lazy]{lazyeval::as.lazy_dots}.}
}
\value{
\code{as.lazy_dots} returns a \link[lazyeval:lazy_dots]{lazyeval::lazy_dots} object.
}
\description{
Convert quotations and dot lists to the representations used
by other packages.
}
