% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dots.R, R/conversions.R
\name{dots}
\alias{dots}
\alias{as.data.frame.dots}
\alias{dots_}
\alias{exprs}
\alias{exprs.dots}
\alias{exprs<-}
\alias{exprs<-.dots}
\alias{envs}
\alias{envs.dots}
\alias{envs<-}
\alias{envs<-.dots}
\alias{[.dots}
\alias{[<-.dots}
\alias{forced.dots}
\alias{c.dots}
\alias{as.dots}
\title{Capture a number of unevaluated arguments as an object.}
\usage{
dots(...)

\method{as.data.frame}{dots}(x)

dots_(exprs, envs)

exprs(x)

\method{exprs}{dots}(d)

exprs(x) <- value

\method{exprs}{dots}(x) <- value

envs(x)

\method{envs}{dots}(x)

envs(x) <- value

\method{envs}{dots}(x) <- value

\method{[}{dots}(x, ..., drop = FALSE)

\method{[}{dots}(x, ...) <- value

\method{forced}{dots}(d)

\method{c}{dots}(...)

as.dots(x)
}
\arguments{
\item{...}{Any number of arguments.}

\item{x}{A \code{\link{dots}} object.}

\item{exprs}{An expression or list of expressions.}

\item{envs}{An environment or list of environments.}

\item{value}{A replacement value.}

\item{x}{A dots object (see \code{\link{dots}}).}

\item{x}{a \code{[dots]} object.}

\item{value}{A replacement list.}

\item{x}{a vector or list.}
}
\value{
A list with class 'dots', each element of which is a \link{quotation}.

\code{as.data.frame.dots} returns a data frame.

\code{dots_(exprs, envs)} directly constructs a dots object
given lists of expresions and environments.

A named list of expressions. The mutator \code{exprs<-}
returns a new dots object with the new expressions.

\code{envs(x)} returns a list of the environments of each
quotation in x.

\code{envs<-} returns a list of new quotations with updated
environments.

\code{envs<-} returns a list of new quotations with updated
expressions.

An object of class \code{\dots}. For \code{as.dots}, the
list items are treated as data values, and create already-forced
promises. For \code{as.dots.exprs}, values are treated as the
expressions of new unforced promises.
}
\description{
A dots object represents a named list of \link{quotation}s. It mirrors R's
special variable \code{...}. Unlike \code{...}, a \code{dots} is:
\itemize{
\item immutable (evaluating does not change it),
\item first-class (you can give it any name, not just \code{...}),
\item data (The R interpreter treates it as literal data rather than
triggering argument splicing).
}

\code{as.data.frame.dots} transforms the contents of a \link{dots} object,
into a data frame with one row per \link{quotation}, with columns:
\itemize{
\item \code{name}: a character,
\item \code{expr}: an expression,
\item \code{env}: an \link{environment} object or NULL if \link{forced},
\item \code{value}: NULL or a value if forced.
}

Access or mutate quotation or dots objects.
\code{exprs(dots(...))} extracts a list of expressions, one per element
of a \code{\link{dots}} object.

\code{envs(dots(...)} extracts a list of environments from a
\link{dots} object.
}
\details{
\code{d <- dots(...)} captures the contents of \code{...} without triggering
evaluation, and returns a list of class \code{"dots"}, each element of
which is a \code{[quotation]}. This improves on
\code{substitute(list(...))[[2]]} by capturing the context of each
expression along with the expressions.

\code{d <- dots(foo, quux=bar+baz)} captures all of the given arguments
in a dots object, like \code{[alist]}, but also captures the
environment of each argument (The present environment in this case).

[[2]: R:[2
[alist]: R:alist
}
\note{
The columns have a class \link{oneline} to better print printing
the data frame.
}
\examples{

named.list <- function(...) {
 # Collect only named arguments, ignoring unnamed arguments.
 d <- dots(...)
 do(list, d[names(d) != ""])
}

named.list(a=1, b=2*2, stop("this is not evaluated"))
}
