% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dots.R
\name{exprs}
\alias{exprs}
\alias{exprs.dots}
\alias{dots_exprs}
\alias{exprs<-}
\alias{envs}
\alias{envs.dots}
\alias{envs<-}
\alias{dots_envs}
\alias{dots_names}
\alias{envs<-.dots}
\alias{dots}
\title{\code{exprs(dots(...))} extracts a list of expressions, one per element
of a \code{\link{dots}} object.}
\usage{
exprs(x)

\method{exprs}{dots}(d)

dots_exprs(...)

exprs(x) <- value

envs(x)

\method{envs}{dots}(x)

envs(x) <- value

dots_envs(...)

dots_names(...)

\method{envs}{dots}(x) <- value

dots(...)
}
\arguments{
\item{x}{A dots object (see \code{\link{dots}}).}

\item{...}{Any number of arguments.}

\item{value}{A list of expressions.}

\item{x}{a \code{[dots]} object.}

\item{value}{A replacement list.}
}
\value{
A named list of expressions. The mutator \code{exprs<-}
returns a new dots object with the new expressions.

\code{dots_exprs(...)} is a shorthand for exprs(dots(...)).

\code{envs(dots(...)} extracts a list of environments from a
\code(link(dots)) object.

\code{envs(x)} returns a list of the environments of each
quotation in x.

\code{envs<-} returns a list of new quotations with updated
environments.

\code{dots_exprs(...)} is a shorthand for `exprs(dots(...))1

\code{dots_names(...)} is a shorthand for \code{names(dots(...))}

\code{envs<-} returns a list of new quotations with updated
expressions.

A list with class 'dots', each element of which is a \link{quotation}.
}
\description{
\code{exprs(dots(...))} extracts a list of expressions, one per element
of a \code{\link{dots}} object.

A dots object represents a named list of quotations. It mirrors R's
special variable \code{...}. Unlike \code{...}, a \code{dots} is:
\itemize{
\item immutable (evaluating does not change it),
\item first-class (you can give it any name, not just \code{...}),
\item data (The R interpreter treates it as literal data rather than
triggering argument splicing).
}
}
\details{
\code{d <- dots(...)} captures the contents of \code{...} without triggering
evaluation, and returns a list of class "dots", each element of
which is a \code{\link{quotation}}. This extends,
e.g. \code{substitute(list(...))[[2]]} by capturing the context of
each expression along with the expressions.

\code{d <- dots(foo, quux=bar+baz)} captures all of its arguments
in a dots object, like \code{\link{alist}()}, but also captures the
environment of each argument.)
}
\examples{

named.list <- function(...) {
 # Collect only named arguments, ignoring unnamed arguments.
 d <- dots(...)
 do(list, d[names(d) != ""])
}

named.list(a=1, b=2*2, stop("this is not evaluated"))
}
