% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/conversions.R
\name{as.dots.environment}
\alias{as.dots.environment}
\alias{dots2env}
\alias{as.environment.dots}
\alias{quo2env}
\title{Capture promises present in an environment and construct a dots object.}
\usage{
\method{as.dots}{environment}(x)

dots2env(d, env = new.env(hash = hash, parent = parent, size = size),
  names = NULL, use_dots = TRUE, append = TRUE, hash = (length(dots) >
  100), size = max(29L, length(dots)), parent = emptyenv())

quo2env(q, env, name)
}
\arguments{
\item{d}{A \link{dots} object.}

\item{env}{Specify an environment object to populate and return. By
default a new environment is created.}

\item{names}{Which variables to populate in the environment. If
NULL is given, will use all names present in the dotlist.  If a
name is given that does not match any names from the dots object,
an error is raised.}

\item{use_dots}{Whether to bind unnamed or unmatched items to
\code{...}. If FALSE, these items are discarded. If TRUE, they
bound to \code{...} in the environment. If items have duplicate
names, the earlier ones are used and the rest placed in "...".}

\item{append}{if \code{TRUE}, unmatched or unnamed items will be
appended to any existing value of '...'. If \code{FALSE}, the existing
binding of \code{...} will be cleared. (Neither happens if \code{use_dots}
is FALSE.)}

\item{hash}{if \code{env} is NULL, this argument is passed to
\code{\link{new.env}}.}

\item{size}{if \code{env} is NULL, this argument is paseed to
\code{\link{new.env}}.}

\item{parent}{if \code{env} is NULL, this argument is paseed to
\code{\link{new.env}}.}

\item{q}{a quotation.}

\item{name}{The name to use. If "" or NULL, will append to "...".}

\item{env}{The environment to store in.}
}
\value{
An environment object.
}
\description{
All bindings in the environment (but not those from its parents) will be
copied into a new \code{\dots} list. Bindings that are promises
will be added to the \dots list without forcing, while other
bindings will be wrapped in an already-forced promise.  If \code{...} is
bound in the environment, all bindings it contains will be added to
the \dots list. The output will not be in any particular order.

All named entries in the dots object will be bound to
variables. Unnamed entries will be appended to any existing value
of \code{...} in the order in which they appear.

\code{quo2env} takes a quotation and creates a single binding in a
given environment.
}
\seealso{
env2dots

env2dots
}
