% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/conversions.R
\name{dots2env}
\alias{dots2env}
\alias{as.environment.dots}
\title{Convert a dots object into promises bound in an envrionment.}
\usage{
dots2env(dots, env = new.env(hash = hash, parent = parent, size = size),
  names = NULL, with_dots = TRUE, hash = (length(dots) > 100),
  size = max(29L, length(dots)), parent = emptyenv())
}
\arguments{
\item{dots}{The dots object to convert.}

\item{names}{Which variables to populate in the
environment. By default, will use all names present in the dotlist.
If a name is given that does not appear in the dots object, an error is
raised.}

\item{with_dots}{Whether to bind unnamed or unmatched items to
\code{...}. If FALSE, these items are discarded. If TRUE, they
are appended to any existing \code{...} in the environment. If
arguments have duplicate names, the earlier ones are used and the
rest placed in "...".}

\item{hash}{if \code{envir} is non-NULL; See \code{\link{new.env}}.}

\item{size}{if \code{envir} is non-NULL; See \code{\link{new.env}}.}

\item{parent}{If creating a new environment, its parent (see
\code{\link{new.env}}).}

\item{envir}{If not NULL, an environment object to populate and
return. If NULL, a new environment will be created.}

\item{parent}{if \code{envir} is non-NULL; See
\code{\link{new.env}}.}
}
\value{
An environment object.
}
\description{
All named entries in the dots object will be bound to
variables. Unnamed entries will be appended to any existing value
of \code{...} in the order in which they appear.
}
