% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/conversions.R
\name{env2dots}
\alias{env2dots}
\title{...}
\usage{
env2dots(env, names = ls(envir = env, all.names = TRUE),
  include_missing = TRUE, expand_dots = TRUE)
}
\arguments{
\item{env}{An environment.}

\item{names}{Which names to extract from the environment. By
default extracts all bindings.}

\item{include_missing}{Whether to include "missing" bindings.}

\item{expand_dots}{Whether to unpack the contents of `...`.}
}
\value{
A \link{dots} object.
}
\description{
...
}
