% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/force.R
\name{forced}
\alias{forced}
\alias{forced_quo}
\alias{forced_quo_}
\alias{forced_dots}
\alias{force_}
\alias{force_.quotation}
\alias{force_.dots}
\alias{forced.quotation}
\alias{forced.default}
\alias{value}
\alias{value.quotation}
\alias{value.dots}
\alias{values}
\alias{values.dots}
\title{Forcing and forcedness of arguments and quotations.}
\usage{
forced(x)

forced_quo(x)

forced_quo_(x)

forced_dots(...)

force_(q, ...)

\method{force_}{quotation}(q, eval = base::eval)

\method{force_}{dots}(d, ...)

\method{forced}{quotation}(x)

\method{forced}{default}(x)

value(q, ...)

\method{value}{quotation}(q, mode = "any")

\method{value}{dots}(d)

values(p)

\method{values}{dots}(d)
}
\arguments{
\item{...}{any number of arguments.}

\item{q}{a \link{quotation} or \link{dots} object.}

\item{eval}{Call something other than 'eval' to "evaluate" the expression.}

\item{mode}{Whether to force in "any" mode or "function" mode (see
\link{locate}).}
}
\value{
\code{forced(x)} returns a \link{logical} value.

\code{forced_quo} and \code{forced_quo_} return \href{quo}{quotation}
objects.

\code{forced_dots} and \code{forced_dots_} return \link{dots} objects.

\code{value} returns the result of forcing the quotation.

\code{values} returns a list.
}
\description{
There are two kinds of \href{quo}{quotations}: forced and unforced.
Unforced quotations have an expression and an environment; forced
quotations have an expression and a value. \code{forced} returns a
logical or logical vector givenm a \link{quotation} or \link{dots}.

\code{forced_quo(x)} constructs a forced quotation directly from its
argument.  It is equivalent to \code{quo(x, force=TRUE)}.

\code{forced_quo_(x)} makes a forced quotation from any data.
Specifically it constructs a \link{quotation} with the same object in
both the \code{expr} and \code{value} slots, except if is a
\href{is.language}{language} object in which case the value is wrapped
in \code{quote()}.

\code{forced_dots} and \code{forced_dots_} construct a \link{dots} object
containing forced quotations.

\code{force_(x)} evaluates the contents of a quotation or dots object.
if they are not already forced, and returns a new forced object.

\code{value} or \code{values} returns the value of a quotation or dots,
forcing it if necessary.
}
\details{
In normal R evaluation, function arguments are recorded as unforced
\link{promise}s, which are converted into forced promises as
well. Quotation objects therefore exist in the same two varieties.
}
\seealso{
is_forced

force
}
