% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/force.R
\name{forced}
\alias{forced}
\alias{forced_quo}
\alias{forced_quo_}
\alias{forced_dots}
\alias{forced_dots_}
\alias{force_}
\alias{force_.quotation}
\alias{force_.dots}
\alias{forced.quotation}
\alias{forced.default}
\alias{value}
\alias{value.quotation}
\alias{values}
\title{Working with forced quotations and dots.}
\usage{
forced(x)

forced_quo(x)

forced_quo_(x)

forced_dots(...)

forced_dots_(values)

force_(q, ...)

\method{force_}{quotation}(q, eval = base::eval)

\method{force_}{dots}(d, ...)

\method{forced}{quotation}(x)

\method{forced}{default}(x)

value(q, ...)

\method{value}{quotation}(q, mode = "any")

values(p)
}
\arguments{
\item{...}{any number of arguments.}

\item{q}{a \link{quotation} or \link{dots} object.}

\item{eval}{Call something other than 'eval' to "evaluate" the expression.}

\item{mode}{Whether to force in "any" mode or "function" mode (see
\link{locate}).}
}
\value{
\code{forced(x)} returns a \link{logical} value.

\code{forced_quo} and \code{forced_quo_} return \href{quo}{quotation} objects.

\code{forced_dots} and \code{forced_dots_} return \link{dots} objects.

\code{value} returns the result of forcing the quotation.

\code{values} returns a list.
}
\description{
In R's implementation of lazy evaluation, function arguments are
first recorded as a \code{promise} which pair of an expression with a
source environment. After a promise's value is needed, it is
converted to a "forced" promise, which records the expression and
value, but no longer records the source environment (this allows
the environments to be garbage collected.) \href{quo}{Quotation} objects
also exist in unforced and forced varieties.

\code{forced_quo(x)} forces its argument before capturing it as a
quotation (including the literal expression). It is equivalent to
\code{quo(x, force=TRUE)}.

\code{forced_quo_(x)} makes a forced quotation from any data.
Specifically it constructs a \link{quotation} with the same
object in both the \code{expr} and \code{value} slots, and does not capture
the expression like \code{forced_quo}.

\code{forced_dots} and \code{forced_dots_} construct a \link{dots} object
containing forced quotations, analogously to \code{forced_quo} and
\code{forced_quo_}.

\code{force_(x)} evaluates the contents of a quotation or dots object,
if they are not already forced, and returns a new, forced quotation
or dots object.

\code{value} or \code{values} returns the value of a quotation, forcing it if
necessary.
}
\seealso{
is_forced

force
}
