% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/format.R
\name{format.oneline}
\alias{format.oneline}
\title{Format a sequence of objects to show one line each.}
\usage{
\method{format}{oneline}(x, width = 30, ...)
}
\arguments{
\item{x}{An object}

\item{width}{the width of line to produce.}

\item{...}{parameters passed to "format"}
}
\description{
Somewhat similar to format.AsIs but tries harder with language objects.
}
