% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/format.R
\name{format.oneline}
\alias{format.oneline}
\title{Formats a sequence of objects to show one line each.
Somewhat similar to format.AsIs but tries harder with language objects.}
\usage{
\method{format}{oneline}(x, width = 30, ...)
}
\arguments{
\item{x}{An object}

\item{width}{}

\item{...}{parameters passed to "format"}
}
\description{
Formats a sequence of objects to show one line each.
Somewhat similar to format.AsIs but tries harder with language objects.
}

