% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/format.R
\name{format.dots}
\alias{format.dots}
\alias{format.quotation}
\alias{format.oneline}
\alias{print.dots}
\alias{print.quotation}
\title{Formatting methods for dots and quotations.}
\usage{
\method{format}{dots}(x, compact = FALSE, show.environments = !compact,
  show.expressions = !compact, width = 30, ...)

\method{format}{quotation}(x, compact = FALSE, show.environments = !compact,
  show.expressions = !compact, width = 30, ...)

\method{format}{oneline}(x, max.width = 50, width = max.width, ...)

\method{print}{dots}(x, ...)

\method{print}{quotation}(x, ...)
}
\arguments{
\item{x}{A dots object.}

\item{compact}{Implies \code{show.environments=FALSE} and
\code{show.expressions=FALSE}.}

\item{show.environments}{Whether to show environments for forced
quotations.}

\item{show.expressions}{Whether to show expressions for unforced
quotations.}

\item{width}{The width of line to produce.}

\item{...}{Further arguments passed to \link{format} methods.}

\item{max.width}{The maximum width of line to produce.}

\item{x}{An object.}

\item{...}{Parameters passed to \link[base:format]{base::format}}
}
\description{
\code{format.dots} constructs a string representation of a dots
object. An unevaluated promise is shown as \code{envir ? expr} and an
evaluated promise is shown as \code{expr := value}.

\code{format.quotation} constructs a string representation of a
quotation object.

The \code{format} method for class \code{oneline} formats a sequence of
objects to show one line each. It is somewhat similar to
\link{format.AsIs} but tries harder with language objects.
}
