% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dots.R
\name{get_dots}
\alias{get_dots}
\title{Inspect "..." in some environment to make sure that If there is no
"..." in the given environment, or it is missing, returns an empty
dotslist.}
\usage{
get_dots(env = caller(environment()))
}
\arguments{
\item{env}{The environment to look in.}
}
\value{
the contents of \code{...} converted to a \code{dots} object.
}
\description{
Inspect "..." in some environment to make sure that If there is no
"..." in the given environment, or it is missing, returns an empty
dotslist.
}
