% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dots.R
\name{get_dots}
\alias{get_dots}
\alias{set_dots}
\title{Set or get the contents of "..." in an environment.}
\usage{
get_dots(env = caller(environment()), inherits = FALSE)

set_dots(env, d, append = FALSE)
}
\arguments{
\item{env}{The environment to look in.}

\item{inherits}{Whether to allow '...' to be found in enclosing
environments.}

\item{d}{a \code{[dots]} object.}

\item{append}{if TRUE, the values should be appended to the
existing binding. If false, existing binding for "..." will be
replaced.}
}
\value{
\code{get_dots} returns the contents of \code{...} converted to a
\code{dots} object.

\code{set_dots} returns the updated environment, invisibly.
}
\description{
Set or get the contents of "..." in an environment.
}
