% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dots.R
\name{get_dots}
\alias{get_dots}
\title{Retrieve the binding of "..." from a given environment.}
\usage{
get_dots(env = caller(environment()), inherits = FALSE)
}
\arguments{
\item{env}{The environment to look in.}

\item{inherits}{Whether to look in enclosing environments for a dotsxp.}
}
\value{
The contents of \code{...} converted to a \code{dots} object.
}
\description{
Retrieve the binding of "..." from a given environment.
}
