% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dots.R
\name{get_dots}
\alias{get_dots}
\title{Retrieve the bindimg of "..." from a given environment.}
\usage{
get_dots(env = caller(environment()))
}
\arguments{
\item{env}{The environment to look in.}
}
\value{
the contents of \code{...} converted to a \code{dots} object.
}
\description{
Retrieve the bindimg of "..." from a given environment.
}
