% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/getpromise.R
\name{locate}
\alias{locate}
\title{Determine which enclosing environments defines a
name.}
\usage{
locate(sym, env = arg_env_(quote(sym), environment()), mode = "any", ...)
}
\arguments{
\item{mode}{Either "any" or "function".}

\item{x}{A name.}
}
\description{
\code{locate} is useful if you want to implement something that works
like \code{<<-}, which updates the binding where it is found.
}
\examples{
`<<-` <- function(lval, rval) {
 lval <- arg(lval)
 rval <- arg(rval)
 target.env <- find_(expr(lval), parent.env(env(lval)))
 do_(quo(`<-`, target.env), lval, arg(rval))
}
}
