% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dots.R
\name{set_dots}
\alias{set_dots}
\title{Set the binding of "..." in an environment.}
\usage{
set_dots(env, d, append = FALSE)
}
\arguments{
\item{env}{The environment to update.}

\item{d}{a \code{\link{dots}} object.}

\item{append}{if TRUE, the values should be appended to the
existing binding. If false, existing binding for "..." will be
replaced.}
}
\value{
The updated environment, invisibly.
}
\description{
If an empty dotlist is given, the "..." binding is cleared.
}
