% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/getpromise.R
\name{unwrap}
\alias{unwrap}
\title{If a quotation is an unevaluated symbol, find the promise through
its source envronments, potentially iterating through several
unforced promises without forcing any. If at any point we come to a
promise with a more-complicated expression, we use that.}
\usage{
unwrap(x)
}
\description{
If something in the promise chain refers to a nonexistent variable,
return an empty promise (i.e. \code{quo()}, which can be tested for using \code{missing_}).
}
\details{
This process is used by the builtin \link{missing}, so this algorithm,
is used by \code{missing_}. It can also be useful to derive axis labels.
}
